harvey-service/requirements.txt
fastapi
uvicorn
openai
langgraph
langchain
pydantic
redis
psycopg2-binary
sqlalchemy
python-dotenv
chromadb
loguru
httpx
python-multipart
aiofiles
pydub
